requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.2

# Data Processing
numpy==1.26.2
pandas==2.1.3
SQLAlchemy==2.0.23

# Web Scraping & HTTP
beautifulsoup4==4.12.2
requests==2.31.0

# Date & Time
python-dateutil==2.8.2

# Image Processing
Pillow==10.1.0

# Markdown Processing
markdown2==2.4.10

# Authentication & Security
PyJWT==2.8.0

# Environment Variables
python-dotenv==1.0.0

# Test Data Generation
Faker==20.1.0